### 2023-11-09 #6039 @editorialbot
[PRE REVIEW]: TDEP: Temperature Dependent Effective Potenials
**Submitting author:** @flokno (<a href="http://orcid.org/0000-0002-7132-039X">Florian Knoop</a>)
**Repository:** https://github.com/tdep-developers/tdep
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
