### 2023-11-09 df7ba0c
Fix for 2D qpoint grid
