+++ src/libolle/type_qpointmesh_gridgeneration.f90
@@ -242,7 +242,7 @@ module subroutine build_grid(qp,p,mw,mem,verbosity,nosymmetry)
             t0=t1
         endif
 
-        
+
 
         ! ! This is an agressive test to make sure all the symmetry operations
         ! ! I identified actually hold. This has not triggered in many years,
@@ -950,12 +950,15 @@ module subroutine tesselate_qgrid(qp,p,mw,mem,verbosity)
 
     ! Build tetrahedrons somehow.
     buildtet: block
+        real(r8), dimension(:,:), allocatable :: center_it,center_at
+        integer, dimension(:,:,:), allocatable :: shift_it,shift_at
         integer, dimension(:,:), allocatable :: ind_it,ind_at
         integer, dimension(:), allocatable :: offset_shell
 
         real(r8), dimension(3,4,6) :: tc0
+        real(r8), dimension(3,6) :: tc1
         real(r8), dimension(3,8) :: bc0
-        real(r8), dimension(3) :: v0,v1,v2,w0
+        real(r8), dimension(3) :: v0,v1,v2,w0,w1
         integer, dimension(4,6) :: tetind,tetjnd
         integer, dimension(8) :: boxind
         integer, dimension(3) :: gi
@@ -969,8 +972,17 @@ module subroutine tesselate_qgrid(qp,p,mw,mem,verbosity)
         nt_tot=qp%n_full_point*6
         call mem%allocate(ind_it,[4,nt_irr],persistent=.false.,scalable=.false.,file=__FILE__,line=__LINE__)
         call mem%allocate(ind_at,[4,nt_tot],persistent=.false.,scalable=.false.,file=__FILE__,line=__LINE__)
+        call mem%allocate(center_it,[3,nt_irr],persistent=.false.,scalable=.false.,file=__FILE__,line=__LINE__)
+        call mem%allocate(center_at,[3,nt_tot],persistent=.false.,scalable=.false.,file=__FILE__,line=__LINE__)
+        call mem%allocate(shift_it,[3,4,nt_irr],persistent=.false.,scalable=.false.,file=__FILE__,line=__LINE__)
+        call mem%allocate(shift_at,[3,4,nt_tot],persistent=.false.,scalable=.false.,file=__FILE__,line=__LINE__)
+
         ind_it=0
         ind_at=0
+        center_it=0.0_r8
+        center_at=0.0_r8
+        shift_it=0
+        shift_at=0
 
         ! And a dummy counter thing
         call mem%allocate(offset_shell,size(prototype_shell),persistent=.false.,scalable=.false.,file=__FILE__,line=__LINE__)
@@ -990,12 +1002,24 @@ module subroutine tesselate_qgrid(qp,p,mw,mem,verbosity)
             ibox=prototype_shell(ii)
             v0=r(:,ibox)
             call neighbour_corners_from_center(qp,v0,boxind,bc0)
-            call chop_box_into_tetrahedrons(p,bc0,tetind,tc0)
+            call chop_box_into_tetrahedrons(p,bc0,tetind,tc0,tc1)
 
             ! Store this irreducible tetrahedron
             do i=1,6
+
                 itet=(ii-1)*6+i
                 ind_it(:,itet)=boxind(tetind(:,i))
+                center_it(:,itet)=tc1(:,i)
+                do j=1,4
+                    ! absolute coordinate of point
+                    w0=qp%ap( ind_it(j,itet) )%r
+                    ! to fractional
+                    w0=matmul(p%inv_reciprocal_latticevectors,w0)
+                    ! distance to box corner
+                    w0=tc0(:,j,i)-w0
+                    ! Store as lattice vector shift
+                    shift_it(:,j,itet)=anint(w0)
+                enddo
[…]
