### 2023-11-09 #6039 @kyleniemeyer
Hi @flokno, I noticed a typo in the paper title: Potenials -> Potentials. Could you fix in the source file?
