### 2023-11-17 #6039 @bjmorgan
@rkurchin I am afraid this wouldn't be possible. I would suggest @alexsquires (or they might be able to suggest other alternatives) who also have more experience than I do with the types of calculations TDEP is intended to help with.
