### 2023-12-04 #6039 @rkurchin
👋 @corettialessandro, @alexsquires, and/or @KEIPERTK, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
