### 2023-12-14 #6150 @editorialbot
[REVIEW]: TDEP: Temperature Dependent Effective Potentials
**Submitting author:** @flokno (<a href="http://orcid.org/0000-0002-7132-039X">Florian Knoop</a>)
**Repository:** https://github.com/tdep-developers/tdep
**Branch with paper.md** (empty if default branch): paper
**Version:** 24.01
