### 2024-01-05 #6150 @flokno
Dear @ejmeitz , thanks for starting the review!

Can you clarify the missing check on `Functionality: Have the functional claims of the software been confirmed?`: Is this a) with respect to the paper, or b) the online documentation incl. tutorials?

For a), we have cited a multitude of peer-reviewed papers that were prepared with TDEP, maybe this was not made clear enough?

For b), we have prepared an extensive set of online tutorials to introduce the core functionality of TDEP in depth. Is there anything else missing?

Another question, maybe also @rkurchin, for `State of the field: Do the authors describe how this software compares to other commonly-used packages?`: We have tried to clarify the _conceptual_ similarity to the most common alternatives in the field (sscha, scp, qscaild) with references -- do you suggest to comment on the _technical_ differences in these approaches? Some hints would be appreciated.

Thank you! And a happy new year to all!
