### 2024-01-05 #6150 @ejmeitz
@flokno I just haven't gotten around to fully testing the package. Hopefully, I'll be done within the week and I'll add another comment with some feedback!

The second check I left empty as I was unsure if what was in the paper was enough. I think Rachel's response should help clarify that.
