### 2024-01-05 #6150 @rkurchin
@flokno, regarding your question about the paper, what you have now looks fine to me – frankly, your paper is already on the longer side for JOSS papers, which tend to be quite brief!
