### 2024-01-07 tdep-developers/tdep#35 @ejmeitz
[ERROR] LAMMPS Format Option Segfaults
When using the option `output_format` on both the `generate_structure` and `canonical_configuration` (and I expect any binary that has this option) I get an error when using the LAMMPS format.

The command I ran was from the MgO [example](https://github.com/tdep-developers/tdep-tutorials/blob/main/02_sampling/example_materials/MgO/README.md), but I added the `-of 3` option to the end:
`canonical_configuration -n 4 -t 300 -mf 20 --quantum -of 3`

```
At line 571 of file type_crystalstructure_io.f90 (unit = 100, file = 'lammps_conf0001')
Fortran runtime error: Expected INTEGER for item 4 in formatted transfer, got REAL
(i10,e20.8,a)
 ^
```
