### 2024-01-07 #6150 @ejmeitz
@flokno Everything looks functional except an issue with the LAMMPS output format. I opened an issue in the main branch:
[#35](https://github.com/tdep-developers/tdep/issues/35)
