### 2024-01-07 c012f0b
output format for lammps was wrong
