### 2024-01-08 #6150 @flokno
> @flokno is there a reason the holding codes are not shipped with the main branch of TDEP? I was hoping to just replace my old install from the TDEP development GitHub with this install, but the holding codes are gone. The build script still has the options for them. Is it safe to just copy the `holding` folder from my old install (I think it is since its a separate binary). I would like to use the `remap_forceconstants` code. If you think its not safe I will mess around with this later as the holding codes are not pertinent to the review.
>
> Edit: It seems to have built fine just copying the holding folder from tdep-devel to tdep.

@ejmeitz indeed, the codes in `holding` (and also `staging`) are more experimental, less well tested, and potentially (though not necessarily) buggy. Also, those features are not needed to use TDEP for the usual use cases as documented. For the public release, we therefore decided to include the mature features only, and add updated versions of codes like `remap_forceconstants` upon popular request. Please feel free to open an issue an request this feature to be added to the public repo (*).

(*): remapping is internally done anyways, so usually `remap_forceconstants` is _not_ explicitly needed for any of the default features of TDEP.
