### 2024-01-08 tdep-developers/tdep#35 @flokno
@mjv500 maybe a good motivation to include https://github.com/tdep-developers/tdep/issues/10 as well (preferably via a PR that closes both issues, but we can do that manually of course)?
