### 2024-01-08 tdep-developers/tdep#35 @mjv500
Good point - I got the message about PRs after having committed. @Ethan -
did you succeed with the new commit?

M.

On Mon, Jan 8, 2024 at 9:35 AM Florian Knoop ***@***.***>
wrote:

> @mjv500 <https://github.com/mjv500> maybe a good motivation to include #10
> <https://github.com/tdep-developers/tdep/issues/10> as well (preferably
> via a PR that closes both issues, but we can do that manually of course)?
>
