### 2024-01-08 tdep-developers/tdep#35 @ejmeitz
The error is now on the next item of the format string:

> At line 571 of file type_crystalstructure_io.f90 (unit = 100, file = 'lammps_conf0001')
> Fortran runtime error: Expected INTEGER for item 5 in formatted transfer, got CHARACTER
> (i10,e20.8,2a)
>  ^
