### 2024-01-09 67a422b
io | fix lammps output
- https://github.com/tdep-developers/tdep/issues/35
