+++ src/libolle/Makefile
@@ -500,6 +500,7 @@ $(OBJECT_PATH)type_crystalstructure_atomdata.o:\
     $(OBJECT_PATH)type_crystalstructure.o
 	$(FC) $(FFLAGS) -c type_crystalstructure_atomdata.f90 -o $@
 $(OBJECT_PATH)type_crystalstructure_io.o:\
+	type_crystalstructure_io.f90\
     $(OBJECT_PATH)konstanter.o\
     $(OBJECT_PATH)hdf5_wrappers.o\
     $(OBJECT_PATH)type_crystalstructure.o
