### 2024-01-09 tdep-developers/tdep#35 @flokno
Ok @ejmeitz this should fix it: https://github.com/tdep-developers/tdep/pull/36

Please confirm on your end and we're good to go
