### 2024-01-09 tdep-developers/tdep#35 @ejmeitz
Command runs well now, I believe you need an new (empty) line after the "Masses" heading for it to get parsed right though. With that space it gets parsed correctly. (quite the annoying format in my opinion). It also prints ` Have not fixed velocity output for LAMMMPS yet` not sure if you meant to leave that there.

Here's an example file that LAMMPS generated internally.
[initial_structure_LJ.txt](https://github.com/tdep-developers/tdep/files/13877922/initial_structure_LJ.txt)
