### 2024-01-10 tdep-developers/tdep#35 @flokno
@ejmeitz ok thanks for checking. LAMMPS seems to be a little problematic.

We discussed the issue and tend towards removing native LAMMPS support as we don't use (and maintain) it ourselves. Converting (default) VASP POSCARs to Lammps is as easy as e.g.

```bash
ase convert contcar_conf0001 -i vasp lammps.in -o lammps-data (--write-args masses=True)
```

Would that be fine?
