### 2024-01-10 tdep-developers/tdep#35 @ejmeitz
Fine by me, I dont actually use the LAMMPS format was just testing stuff for the JOSS paper and this was the thing that broke.

That format is problematic in many ways, I understand why this would be faster.
