### 2024-01-10 c9dcab6
io | remove LAMMPS writer
- remove from crystalstructure_io
- remove from programs
- remove from docs
