### 2024-01-10 2dadb8e
build_things | remove holding from public version
