### 2024-01-10 tdep-developers/tdep#35 @flokno
@ejmeitz ok I removed lammps altogether now. I'll update the docs once we all agree on this.

I also removed the `holding` stuff from `build_things.sh`
