### 2024-01-10 f2fab5d
BUG: Fix IO for LAMMPS and Siesta (#36)
* io | fix lammps output

- https://github.com/tdep-developers/tdep/issues/35

* io | fix Siesta output

- https://github.com/tdep-developers/tdep/issues/10

* io | remove LAMMPS writer

- remove from crystalstructure_io
- remove from programs
- remove from docs

* build_things | remove holding from public version
