+++ src/libolle/type_crystalstructure_io.f90
@@ -515,71 +515,73 @@ module subroutine writetofile(p,filename,output_format,write_velocities,transfor
         logical :: writevel
         integer :: i,u,it
 
-        if ( present(write_velocities) ) then
-            writevel=write_velocities
-        else
-            writevel=.false.
-        endif
+        call lo_stop_gracefully(['Native LAMMPS IO was removed, please use external converters.'], 8)
 
-        ! Get the lattice parameter thingy
-        call lo_get_axis_angles(p%latticevectors,a,b,c,al,be,gm)
+        ! if ( present(write_velocities) ) then
+        !     writevel=write_velocities
+        ! else
+        !     writevel=.false.
+        ! endif
 
-        ax=a
-        bx=lo_chop(b*cos(gm),lo_sqtol)
-        by=sqrt(b**2-bx**2)
-        cx=lo_chop( c*cos(be) ,lo_sqtol)
-        cy=lo_chop( ( dot_product(p%latticevectors(2,:),p%latticevectors(3,:))-bx*cx )/by ,lo_sqtol)
-        cz=lo_chop( sqrt(c**2-cx**2-cy**2) , lo_sqtol )
+        ! ! Get the lattice parameter thingy
+        ! call lo_get_axis_angles(p%latticevectors,a,b,c,al,be,gm)
 
-        ! Convert to Angstrom
-        ax=ax*lo_bohr_to_A
-        bx=bx*lo_bohr_to_A
-        by=by*lo_bohr_to_A
-        cx=cx*lo_bohr_to_A
-        cy=cy*lo_bohr_to_A
-        cz=cz*lo_bohr_to_A
+        ! ax=a
+        ! bx=lo_chop(b*cos(gm),lo_sqtol)
+        ! by=sqrt(b**2-bx**2)
+        ! cx=lo_chop( c*cos(be) ,lo_sqtol)
+        ! cy=lo_chop( ( dot_product(p%latticevectors(2,:),p%latticevectors(3,:))-bx*cx )/by ,lo_sqtol)
+        ! cz=lo_chop( sqrt(c**2-cx**2-cy**2) , lo_sqtol )
 
-        basis=0.0_flyt
-        basis(:,1)=[ ax , 0.0_flyt, 0.0_flyt]
-        basis(:,2)=[ bx, by, 0.0_flyt]
-        basis(:,3)=[ cx, cy, cz ]
-        basis=lo_chop(basis,lo_sqtol)
+        ! ! Convert to Angstrom
+        ! ax=ax*lo_bohr_to_A
+        ! bx=bx*lo_bohr_to_A
+        ! by=by*lo_bohr_to_A
+        ! cx=cx*lo_bohr_to_A
+        ! cy=cy*lo_bohr_to_A
+        ! cz=cz*lo_bohr_to_A
 
-        ! Return the coordinate transformation to what LAMMPS likes.
-        if ( present(transformationmatrix) ) then
-            transformationmatrix=matmul(basis,p%inv_latticevectors)
-        endif
+        ! basis=0.0_flyt
+        ! basis(:,1)=[ ax , 0.0_flyt, 0.0_flyt]
+        ! basis(:,2)=[ bx, by, 0.0_flyt]
+        ! basis(:,3)=[ cx, cy, cz ]
+        ! basis=lo_chop(basis,lo_sqtol)
 
-        ! figure out how to convert stuff
-        tm=matmul(basis,p%inv_latticevectors)
+        ! ! Return the coordinate transformation to what LAMMPS likes.
+        ! if ( present(transformationmatrix) ) then
+        !     transformationmatrix=matmul(basis,p%inv_latticevectors)
+        ! endif
 
-        ! and print it
-        u=open_file('out',trim(filename))
[…]
