### 2024-01-10 #6150 @flokno
> @flokno thanks for dealing with the LAMMPS stuff. I checked off the last box on my review!
>
> I do have a question about the remap_forceconstant still. TDEP gives the force constants for a unit cell of the material; however, my calculations require the full force constant matrix for a supercell. Olle had told me to use remap_forceconstant for this. Is there still a way to get the supercell forceconstants from TDEP or will I have to write my own script? (can move to the discussion board on TDEP if needed)

Ok I understand. For the time being you can just use the code in `tdep-devel`, the code base is basically the same. If you think it should be added to the public repo, please request the feature per issue.

**Thank you for the review and the feedback!**

@rkurchin and @corettialessandro , any more comments?
