### 2024-01-16 816607b
extract_forceconstants | dump energies to file for --verbose
