### 2024-01-17 9b93ac7
extract_forceconstants | dump energies in eV/atom
* add infos to `outfile.U0`
