+++ src/extract_forceconstants/main.f90
@@ -442,7 +442,7 @@ getU0: block
 
     ! file for the energies
     u = open_file('out', 'outfile.energies')
-    write (u, '(A,A)') '# Unit:      ', 'meV/atom'
+    write (u, '(A,A)') '# Unit:      ', 'eV/atom'
     write (u, '(A,A)') '# no. atoms: ', tochar(ss%na)
     write (u, "(A)") '#  conf    Epot                  Epolar                &
         &Epair                 Etriplet              Equartet'
@@ -546,7 +546,7 @@ getU0: block
         if ((mw%talk) .and. (opts%verbosity > 0)) then
             ctr = ctr + 1
             ! Dump it to file
-            write (u, "(1X,I5,5(2X,E20.12))") t, e0(t)*tomev, ep(t)*tomev, e2(t)*tomev, e3(t)*tomev, e4(t)*tomev
+            write (u, "(1X,I5,5(2X,E20.12))") t, e0(t)*toev, ep(t)*toev, e2(t)*toev, e3(t)*toev, e4(t)*toev
 
             if (lo_trueNtimes(ctr, 20, ctrtot)) then
                 write (*, "(1X,I5,F30.12,4(2X,F20.12))") t, e0(t)*tomev, ep(t)*tomev, e2(t)*tomev, e3(t)*tomev, e4(t)*tomev
@@ -626,7 +626,12 @@ getU0: block
 
         ! Dump it to file
         u = open_file('out', 'outfile.U0')
-        write (u, "(4(1X,E19.12))") &
+        write (u, '(A,A)') '# Unit:      ', 'eV/atom'
+        write (u, '(A,A)') '# no. atoms: ', tochar(ss%na)
+        write (u, "(A)") '#            mean(Epot)                     mean(Epot - Epolar - E2) &
+            &      mean(Epot - Epolar - E2 - E3)  mean(Epot - Epolar - E2 - E3 - E4)'
+
+        write (u, "(4(1X,E30.12))") &
             (baseline + lo_mean(e0))*toev, &
             (baseline + lo_mean(e0 - ep - e2))*toev, &
             (baseline + lo_mean(e0 - ep - e2 - e3))*toev, &
