### 2024-01-17 #6150 @corettialessandro
I am still checking some functionalities via the tutorials, I should be done for the end of the week.

In the meantime, in the file `important_settings.conda` the default PATH is still the one from one of the authors. You should maybe think about changing it to something more neutral like `path/to/your/conda/environment/` or something like that.

Running the test suite goes without issues, it only yields a warning that I was not able to address:

```
atomic_distribution/test_atomic_distribution.py::test_hdf5
  <frozen importlib._bootstrap>:241: RuntimeWarning: numpy.ndarray size changed, may indicate binary incompatibility. Expected 16 from C header, got 96 from PyObject

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
```
If it is something that also others are experiencing, maybe you could  mention it in the README in the tests folder.

Finally, I would remove the self-link from the webpage of the manual. I understand that it is arising from the markdown README of GitHub, but it gives the impression that there is yet another page for the manual, and the link redirects to the same page. I would simply remove the section `Manual` from the webpage.

@rkurchin these are not serious problems that prevent publication, should I still open issues about them?
