### 2024-01-17 a2680cf
important_settings | update conda template
