### 2024-01-17 #6150 @flokno
Hi @corettialessandro ,

- ok for the path, addressed in https://github.com/tdep-developers/tdep/commit/a2680cf22aaa7d949b3c93da79f6577cd160a976
- The warning is about your python environment, this typically occurs when a different `numpy` version was used when building the environment vs. when running the application. I think it's a good idea to mention those in the readme, done in https://github.com/tdep-developers/tdep/commit/ea40439fdd80cbbb13f64a3f523380a44b3e6dd7
- I will see if there is a good solution for the link to the manual.

Thanks!
