### 2024-01-19 0761b03
fprettifying anharmonic free energy
