+++ src/anharmonic_free_energy/options.f90
@@ -1,7 +1,7 @@
 #include "precompilerdefinitions"
 module options
-use konstanter, only: r8,lo_author,lo_version,lo_licence,lo_tiny,lo_status,lo_exitcode_baddim,lo_exitcode_param,&
-                      lo_frequency_Hartree_to_THz,lo_frequency_Hartree_to_icm,lo_frequency_Hartree_to_meV
+use konstanter, only: r8, lo_author, lo_version, lo_licence, lo_tiny, lo_status, lo_exitcode_baddim, lo_exitcode_param, &
+                      lo_frequency_Hartree_to_THz, lo_frequency_Hartree_to_icm, lo_frequency_Hartree_to_meV
 use gottochblandat, only: lo_stop_gracefully
 use flap, only: command_line_interface
 implicit none
@@ -16,8 +16,8 @@ type lo_opts
     logical :: stochastic = .false.
     logical :: thirdorder = .false.
     logical :: fourthorder = .false.
-    contains
-        procedure :: parse
+contains
+    procedure :: parse
 end type
 
 contains
@@ -33,62 +33,62 @@ subroutine parse(opts)
     integer :: errctr
 
     ! basic info
-    call cli%init(progname    = 'anharmonic_free_energy',&
-              authors     = lo_author,&
-              version     = lo_version,&
-              license     = lo_licence,&
-              help        = 'Usage: ',&
-              description = 'Calculates the anharmonic free energy.',&
-              examples    = ["mpirun anharmonic_free_energy"],&
-              epilog      = new_line('a')//"...")
+    call cli%init(progname='anharmonic_free_energy', &
+                  authors=lo_author, &
+                  version=lo_version, &
+                  license=lo_licence, &
+                  help='Usage: ', &
+                  description='Calculates the anharmonic free energy.', &
+                  examples=["mpirun anharmonic_free_energy"], &
+                  epilog=new_line('a')//"...")
 
     ! Specify some options
     cli_qpoint_grid
     call cli%add(switch='--quantum', &
-        help='Use Bose-Einstein occupations to compute the free energy', &
-        required=.false., act='store_true',def='.false.',error=lo_status)
-        if ( lo_status .ne. 0 ) stop
+                 help='Use Bose-Einstein occupations to compute the free energy', &
+                 required=.false., act='store_true', def='.false.', error=lo_status)
+    if (lo_status .ne. 0) stop
     call cli%add(switch='--stochastic', &
-        help='Add second order cumulant contribution to the free energy with a minus sign for self-consistent sampling', &
-        required=.false., act='store_true',def='.false.',error=lo_status)
-        if ( lo_status .ne. 0 ) stop
+                 help='Add second order cumulant contribution to the free energy with a minus sign for self-consistent sampling', &
+                 required=.false., act='store_true', def='.false.', error=lo_status)
+    if (lo_status .ne. 0) stop
     call cli%add(switch='--thirdorder', &
-        help='Compute third order anharmonic correction to the free energy', &
-        required=.false., act='store_true',def='.false.',error=lo_status)
+                 help='Compute third order anharmonic correction to the free energy', &
+                 required=.false., act='store_true', def='.false.', error=lo_status)
     if (lo_status .ne. 0) stop
     call cli%add(switch='--fourthorder', &
-        help='Compute fourth order anharmonic correction to the free energy', &
-        required=.false., act='store_true',def='.false.',error=lo_status)
[…]
