+++ src/anharmonic_free_energy/main.f90
@@ -1,8 +1,8 @@
 program anharmonic_free_energy
 !!{!src/anharmonic_free_energy/manual.md!}
-use konstanter, only: r8,lo_Hartree_to_eV,lo_kb_Hartree,lo_pressure_HartreeBohr_to_GPa
-use gottochblandat, only: open_file,walltime,lo_linspace,lo_progressbar_init,lo_progressbar,tochar,&
-    lo_does_file_exist,lo_mean,lo_stddev
+use konstanter, only: r8, lo_Hartree_to_eV, lo_kb_Hartree, lo_pressure_HartreeBohr_to_GPa
+use gottochblandat, only: open_file, walltime, lo_linspace, lo_progressbar_init, lo_progressbar, tochar, &
+                          lo_does_file_exist, lo_mean, lo_stddev
 use mpi_wrappers, only: lo_mpi_helper
 use lo_memtracker, only: lo_mem_helper
 use lo_timetracker, only: lo_timer
@@ -11,7 +11,7 @@ use type_forceconstant_secondorder, only: lo_forceconstant_secondorder
 use type_forceconstant_thirdorder, only: lo_forceconstant_thirdorder
 use type_forceconstant_fourthorder, only: lo_forceconstant_fourthorder
 use type_mdsim, only: lo_mdsim
-use type_qpointmesh, only: lo_qpoint_mesh,lo_generate_qmesh,lo_read_qmesh_from_file,lo_fft_mesh
+use type_qpointmesh, only: lo_qpoint_mesh, lo_generate_qmesh, lo_read_qmesh_from_file, lo_fft_mesh
 use type_phonon_dispersions, only: lo_phonon_dispersions
 use lo_phonon_bandstructure_on_path, only: lo_phonon_bandstructure
 use type_phonon_dos, only: lo_phonon_dos
@@ -34,60 +34,60 @@ type(lo_mem_helper) :: mem
 type(lo_timer) :: tmr
 type(lo_mdsim) :: sim
 
-real(r8), dimension(3,5) :: cumulant
-real(r8) :: timer_init,timer_total
-real(r8) :: U0,U1,energy_unit_factor
+real(r8), dimension(3, 5) :: cumulant
+real(r8) :: timer_init, timer_total
+real(r8) :: U0, U1, energy_unit_factor
 logical :: havehighorder
 
 ! Init MPI, timers and options
 call mw%init()
-timer_total=walltime()
-timer_init=walltime()
+timer_total = walltime()
+timer_init = walltime()
 call opts%parse()
 call tmr%start()
 
 ! set up all possible meshes and get all dispersions
 init: block
-    integer :: t,u
+    integer :: t, u
 
     ! only be verbose on the first rank
-    if ( .not. mw%talk ) opts%verbosity=-10
+    if (.not. mw%talk) opts%verbosity = -10
 
     ! Read structure
-    call uc%readfromfile('infile.ucposcar',verbosity=opts%verbosity)
-    call uc%classify('wedge',timereversal=.true.)
-    if ( mw%talk ) write(*,*) '... read unitcell'
+    call uc%readfromfile('infile.ucposcar', verbosity=opts%verbosity)
+    call uc%classify('wedge', timereversal=.true.)
+    if (mw%talk) write (*, *) '... read unitcell'
 
     ! Read forceconstants
-    call fc%readfromfile(uc,'infile.forceconstant',mem,verbosity=-1)
-    if ( mw%talk ) write(*,*) '... read second order forceconstant'
+    call fc%readfromfile(uc, 'infile.forceconstant', mem, verbosity=-1)
[…]

[…]
