### 2024-01-19 4f9fddb
Adding classical limit in anharmonic free energy (#18)
* correcting occupation in third order free energy and adding possibility of classical limit

* changing --classical option to --quantum, correcting the per atom normalization in the output and changing the output of the test for the new default behavior

* Correcting two bugs in anharmonic_free_energy : 1) e4 that had a plus instead of a minus sign 2) use of total energy instead of potential for the computation of U0

* Adding second order cumulant according to the type of sampling

* Add options for third and fourth order, redid the outp
[…]
