### 2024-01-19 #6150 @corettialessandro
Hi @flokno,

Everything fine here, this repository is really impressive!
Nothing really important to say, I checked all the boxes in the review list.

Minor points:

- The routine `extract_forceconstants` can be run with the option `--polar` only if an input file `infile.lotosplitting` is present in the folder. If this is not the case the error message is really cryptic. I guess it should be easy to make it more informative. I also imagine that this could be the case for other routines (if it is not even better). In future versions some improvements in handling of errors would benefit user and developers.
- In the tutorial page for sTDEP with DFT forces there are some typos in the README. It is not important in the text, but it could become tedious when the routines are misspelled (e.g. `generature_structure` or `extract_forceconsant). Take a look at that when you have time.
