### 2024-01-19 #6150 @flokno
Thanks @corettialessandro !

You're right with the error messages. I'll see if I get around this.

If you remember which typos you stumbled over, I'd appreciate if you open an issue with the respective line(s).
