### 2024-01-19 #6150 @rkurchin
Editorial comments:
* line 30: add a comma after "for instance"
* 44: hyphenate "lowest-order" ([compound adjectives](https://www.grammarly.com/blog/compound-adjectives/))
* 53: similarly, hyphenate "lattice-dynamics"
* 55-56: "periodically-arranged"
* 60: add a comma after "problem"
* 65, 97: remove the "Ref." (not necessary due to the way citations are typeset)
* 72: add hyphen after "_non_"
* 98: "light-scattering"
* 101: "open-source"
* 102: either capitalize the spelling-out of MPI or just remove it (IMO it's sufficiently widely known)
* 108: remove comma after "set"
* 119: "code-agnostic" and "plain-text" (or as one word, plaintext)
* 120: capitalize "CSV"
* 130: "code" twice in the sentence is awkward, either change the first one to "functions" (or whatever is appropriate) or maybe just remove the second one and make it "shipped with TDEP"
* 142: add comma after "Optionally"
* 156: capitalize "Atomic Simulation Environment"

References:
* there are some inconsistencies in the way that they're rendering (e.g. sometimes with first name/initial and sometimes not, and I can't discern the origin of that behavior in your .bib file so I'm going to tag in @openjournals/pe-eics on that...
* remove the extra sets of `{}` around titles so that they typeset in sentence case (this applies to a number of references)
* in `Boer.1948`: surname should be "de Boer", I would also change the title to `Quantum theory of condensed permanent gases I: the law of corresponding states` to make it look a little less weird
* in `Born.1912`: second surname should be "von Karman"
* in `Dove.1986`: "Lynden-bell" -> "Lynden-Bell"
