### 2024-01-19 #6150 @kyleniemeyer
Regarding the first name vs first initial question, I think pandoc distinguishes between references where one has the name spelled out and the other just has the initials. In other words, it thinks they may be different people. To avoid this, I recommend cleaning up the `.bib` file to just include first and middle initials.
