### 2024-01-22 481394f
paper | finalize literature
* reformat names
