+++ paper/literature.bib
@@ -120,7 +120,7 @@
   year     = {2021},
 }
 
-@article{Dangić.2021,
+@article{Dangic.2021,
   abstract = {{The proximity to structural phase transitions in IV-VI thermoelectric materials is one of the main reasons for their large phonon anharmonicity and intrinsically low lattice thermal conductivity κ. However, the κ of GeTe increases at the ferroelectric phase transition near 700 K. Using first-principles calculations with the temperature dependent effective potential method, we show that this rise in κ is the consequence of negative thermal expansion in the rhombohedral phase and increase in the phonon lifetimes in the high-symmetry phase. Strong anharmonicity near the phase transition induces non-Lorentzian shapes of the phonon power spectra. To account for these effects, we implement a method of calculating κ based on the Green-Kubo approach and find that the Boltzmann transport equation underestimates κ near the phase transition. Our findings elucidate the influence of structural phase transitions on κ and provide guidance for design of better thermoelectric materials.}},
   author   = {Dangić, Đorđe and Hellman, Olle and Fahy, Stephen and Savić, Ivana},
   doi      = {10.1038/s41524-021-00523-7},
