+++ paper/literature_final.bib
@@ -1,0 +1,737 @@
+@misc{tdeptools,
+  author    = {Knoop, F.},
+  publisher = {GitHub},
+  url       = {https://github.com/flokno/tools.tdep},
+  title     = {{tdeptools: tools for TDEP}},
+  year      = {2023},
+}
+
+@article{Castellano.2023,
+  author   = {Castellano, A. and Batista, J. and Verstraete, M. J.},
+  doi      = {10.48550/arxiv.2303.10621},
+  eprint   = {2303.10621},
+  journal  = {arXiv},
+  title    = {{Mode-coupling theory of lattice dynamics for classical and quantum crystals}},
+  year     = {2023},
+}
+
+@article{Reig.2022,
+  author   = {Reig, D. S. and Varghese, S. and Farris, R. and Block, A. and Mehew, J. D. and Hellman, O. and Woźniak, P. and Sledzinska, M. and Sachat, A. E. and Chávez‐Ángel, E. and Valenzuela, S. O. and Hulst, N. F. van and Ordejón, P. and Zanolli, Z. and Torres, C. M. S. and Verstraete, M. J. and Tielrooij, K.‐J.},
+  doi      = {10.1002/adma.202108352},
+  issn     = {0935-9648},
+  journal  = {Advanced Materials},
+  number   = {10},
+  pages    = {2108352},
+  title    = {{Unraveling Heat Transport and Dissipation in Suspended MoSe2 from Bulk to Monolayer}},
+  volume   = {34},
+  year     = {2022},
+}
+
+@article{Menahem.2022,
+  author   = {Menahem, M. and Benshalom, N. and Asher, M. and Aharon, S. and Korobko, R. and Safran, S. and Hellman, O. and Yaffe, O.},
+  eprint   = {2208.05563},
+  journal  = {arXiv},
+  title    = {{The Disorder Origin of Raman Scattering In Perovskites Single Crystals}},
+  year     = {2022},
+}
+
+@article{Benshalom.2022v0s,
+  author   = {Benshalom, N. and Asher, M. and Jouclas, R. and Korobko, R. and Schweicher, G. and Liu, J. and Geerts, Y. and Hellman, O. and Yaffe, O.},
+  eprint   = {2204.12528},
+  journal  = {arXiv},
+  title    = {{Phonon-phonon interactions in the polarization dependence of Raman scattering}},
+  year     = {2022},
+}
+
+@article{Laniel.2022,
+  author   = {Laniel, D. and Trybel, F. and Winkler, B. and Knoop, F. and Fedotenko, T. and Khandarkhaeva, S. and Aslandukova, A. and Meier, T. and Chariton, S. and Glazyrin, K. and Milman, V. and Prakapenka, V. and Abrikosov, I. A. and Dubrovinsky, L. and Dubrovinskaia, N.},
+  doi      = {10.1038/s41467-022-34755-y},
+  eprint   = {2208.10418},
+  journal  = {Nature Communications},
+  keywords = {me},
+  number   = {1},
+  pages    = {6987},
+  title    = {{High-pressure synthesis of seven lanthanum hydrides with a significant variability of hydrogen content}},
+  volume   = {13},
+  year     = {2022},
+}
+
+@article{Lin.2022,
+  author   = {Lin, C. and Poncé, S. and Marzari, N.},
+  doi      = {10.1038/s41524-022-00920-6},
+  eprint   = {2209.09520},
+  journal  = {npj Computational Materials},
+  number   = {1},
+  pages    = {236},
+  title    = {{General invariance and equilibrium conditions for lattice dynamics in 1D, 2D, and 3D materials}},
+  volume   = {8},
+  year     = {2022},
+}
+
+@article{Cohen.2022,
[…]
