+++ paper/literature_final.bib
@@ -694,7 +694,7 @@
   journal  = {Zeitschrift für Physik},
   number   = {1},
   pages    = {42--57},
-  title    = {Anharmonische Gitterschwingungen und die lineare Kette},
+  title    = {{Anharmonische Gitterschwingungen und die lineare Kette}},
   volume   = {142},
   year     = {1955},
 }
@@ -703,34 +703,35 @@
   author    = {Born, M. and Huang, K.},
   location  = {Oxford},
   publisher = {Clarendon Press},
-  title     = {Dynamical theory of crystal lattices},
+  title     = {{Dynamical theory of crystal lattices}},
   year      = {1954},
+  doi       = {10.1093/oso/9780192670083.001.0001},
 }
 
 @article{Born.1951,
   author = {Born, M. and Brix, P. and Kopfermann, H. and Heisenberg, W. and Staudinger, H. and Stille, H. and Weizsäcker, C. F. v. and Euler, H. v. and Hedvall, J. A. and Siegel, C. L. and Rellich, F. and Nevanlinna, R.},
   doi    = {10.1007/978-3-642-86703-3},
-  title  = {Festschrift zur Feier des Zweihundertjährigen Bestehens der Akademie der Wissenschaften in Göttingen, I. Mathematisch-Physikalische Klasse},
+  title  = {{Festschrift zur Feier des Zweihundertjährigen Bestehens der Akademie der Wissenschaften in Göttingen, I. Mathematisch-Physikalische Klasse}},
   year   = {1951},
 }
 
 @article{Boer.1948,
-  author   = {De Boer, J.},
+  author   = {de Boer, J.},
   doi      = {10.1016/0031-8914(48)90032-9},
   issn     = {0031-8914},
   journal  = {Physica},
   number   = {2-3},
   pages    = {139--148},
-  title    = {Quantum theory of condensed permanent gases I: the law of corresponding states},
+  title    = {{Quantum theory of condensed permanent gases I: the law of corresponding states}},
   volume   = {14},
   year     = {1948},
 }
 
 @article{Born.1912,
-  author  = {Born, M. and Karman, T. v.},
+  author  = {Born, M. and von Karman, T.},
   journal = {Physikalische Zeitschrift},
   pages   = {297--309},
-  title   = {Über Schwingungen in Raumgittern},
+  title   = {{Über Schwingungen in Raumgittern}},
   volume  = {13},
   year    = {1912},
 }
