+++ paper/paper.md
@@ -88,7 +88,7 @@ The chemical bonding in the lattice dynamics Hamiltonian is represented through 
 
 While the theoretical foundation of (self-consistent) phonon theory has been well-established for decades, more recent developments are concerned with implementing this theory in computer simulations, typically based on density functional theory (DFT) [@Hohenberg.1964; @Kohn.1965]. This has led to a variety of approaches that tackle the self-consistent phonon problem for anharmonic and dynamically stabilized systems [@Souvatzis.2008; @Errea.2013; @Tadano.2015; @Roekeghem.2021; @Monacelli.2021]. Another development was the _non_ self-consistent construction of effective Hamiltonians by optimizing the force constants to the fully anharmonic dynamics observed during MD simulations [@Levy.1984; @Dove.1986]. This idea was extended in the TDEP method to describe phonons in systems like Zirconium in the high-temperature body-centered cubic (bcc) phase, which is dynamically unstable at low temperatures [@Hellman.2011; @Hellman.2013]. While the initial TDEP method was based on _ab initio_ MD, a self-consistent extension was later proposed in the form of _stochastic_ TDEP (sTDEP), where thermal samples are created from the model Hamiltonian itself and the force constants are optimized iteratively until self-consistency [@Shulumba.2017; @Benshalom.2022]. sTDEP furthermore allows to include nuclear quantum effects in materials with light elements in a straightforward way [@Shulumba.20179s8e; @Laniel.2022].
 
-Effective phonons capture anharmonic frequency renormalization, but they are still non-interacting quasiparticles with infinite lifetime, or equivalently infinitesimal linewidth. The effect of linewidth broadening due to anharmonic phonon-phonon interactions can be included by using higher-order force constants up to third or fourth order [@Cowley.1963; @Hellman.2013oi5; @Feng.2016]. These can be used to get better approximations to the free energy [@Wallace.1972], describe thermal transport [@Broido.2007; @Romero.2015; @Klarbring.2020vk; @Dangic.2021], and linewidth broadening in spectroscopic experiments [@Romero.2015; @Kim.2018; @Benshalom.2022]. In practice, it was noted that the renormalized phonon quasiparticles interact more weakly than bare phonons. This means that the effective approach remains applicable in systems with strong anharmonicity where the bare phonon quasiparticle picture becomes invalid [@Ravichandran.2018]. A formal justification in terms of mode-coupling theory, as well as a detailed comparison between bare perturbation theory with force constants from a Taylor expansion, self-consistent effective, and non-self-consistent effective approaches was recently given by some of us in Ref. [@Castellano.2023]. Explicitly incorporating dielectric response properties for light scattering experiments such as infrared and Raman was recently proposed [@Benshalom.2022].
[…]
