+++ paper/literature_final.bib
@@ -7,12 +7,17 @@
 }
 
 @article{Castellano.2023,
-  author   = {Castellano, A. and Batista, J. and Verstraete, M. J.},
-  doi      = {10.48550/arxiv.2303.10621},
-  eprint   = {2303.10621},
-  journal  = {arXiv},
-  title    = {{Mode-coupling theory of lattice dynamics for classical and quantum crystals}},
-  year     = {2023},
+    author  = {Castellano, Aloïs and Batista, J. P. Alvarinhas and Verstraete, Matthieu J.},
+    title   = "{Mode-coupling theory of lattice dynamics for classical and quantum crystals}",
+    journal = {The Journal of Chemical Physics},
+    volume  = {159},
+    number  = {23},
+    pages   = {234501},
+    year    = {2023},
+    month   = {12},
+    issn    = {0021-9606},
+    doi     = {10.1063/5.0174255},
+    url     = {https://doi.org/10.1063/5.0174255},
 }
 
 @article{Reig.2022,
@@ -716,7 +721,7 @@
 }
 
 @article{Boer.1948,
-  author   = {de Boer, J.},
+  author   = {{de Boer}, J.},
   doi      = {10.1016/0031-8914(48)90032-9},
   issn     = {0031-8914},
   journal  = {Physica},
@@ -728,7 +733,7 @@
 }
 
 @article{Born.1912,
-  author  = {Born, M. and von Karman, T.},
+  author  = {Born, M. and {von Karman}, T.},
   journal = {Physikalische Zeitschrift},
   pages   = {297--309},
   title   = {{Über Schwingungen in Raumgittern}},
