+++ paper/paper.md
@@ -78,37 +78,37 @@ The Temperature Dependent Effective Potential (TDEP) method is a versatile and e
 
 # Introduction
 
-The properties of materials change both qualitatively and quantitatively with temperature, i.e., the macroscopic manifestation of the microscopic vibrational motion of electrons and nuclei. In thermal equilibrium, temperature influences the structural phase, the density, and many mechanical properties. Out of thermal equilibrium, for instance when applying a thermal gradient or an external spectroscopic probe such as a light or neutron beam, temperature influences the response of the material to the perturbation, for example its ability to conduct heat, or the lineshape of the spectroscopic signal. Temperature is therefore at the core of both applied and fundamental materials science.
+The properties of materials change both qualitatively and quantitatively with temperature, i.e., the macroscopic manifestation of the microscopic vibrational motion of electrons and nuclei. In thermal equilibrium, temperature influences the structural phase, the density, and many mechanical properties. Out of thermal equilibrium, for instance, when applying a thermal gradient or an external spectroscopic probe such as a light or neutron beam, temperature influences the response of the material to the perturbation, for example its ability to conduct heat, or the lineshape of the spectroscopic signal. Temperature is therefore at the core of both applied and fundamental materials science.
 
 In _ab initio_ materials modeling, the contribution of electronic temperature is straightforward to include through appropriate occupation of the electronic states, whereas the nuclear contribution needs to be accounted for explicitly. This can be done by performing molecular dynamics (MD) simulations which aim at _numerically_ reproducing the thermal nuclear motion in an atomistic simulation, and obtaining temperature-dependent observables either in equilibrium, through averaging, or out of equilibrium from time-dependent correlation functions or by directly observing the relaxation dynamics.
 
-An alternative strategy is to construct approximate model Hamiltonians for the nuclear subsystem, which enables an _analytic_ description of the nuclear motion by leveraging perturbation theory starting from an exactly solvable lowest order model. In this approach, the starting point is given in terms of _phonons_ which are eigensolutions of a Hamiltonian of harmonic form, i.e., quadratic in the nuclear displacements. Anharmonic contributions can be included via established perturbative techniques, in practice up to quartic terms. Higher-order contributions are elusive because the complexity and number of terms in the perturbative expansion grows very quickly with system size and perturbation order. The lattice dynamics approach is therefore not formally exact. However, the phonon picture is useful for describing a wide range of materials properties in practice, and often reaches excellent accuracy in comparison to experiment while providing precise microscopic insight into the underlying physical phenomena.
+An alternative strategy is to construct approximate model Hamiltonians for the nuclear subsystem, which enables an _analytic_ description of the nuclear motion by leveraging perturbation theory starting from an exactly solvable lowest-order model. In this approach, the starting point is given in terms of _phonons_ which are eigensolutions of a Hamiltonian of harmonic form, i.e., quadratic in the nuclear displacements. Anharmonic contributions can be included via established perturbative techniques, in practice up to quartic terms. Higher-order contributions are elusive because the complexity and number of terms in the perturbative expansion grows very quickly with system size and perturbation order. The lattice-dynamics approach is therefore not formally exact. However, the phonon picture is useful for describing a wide range of materials properties in practice, and often reaches excellent accuracy in comparison to experiment while providing precise microscopic insight into the underlying physical phenomena.
 
[…]
