+++ paper/literature_final.bib
@@ -534,7 +534,7 @@
 }
 
 @article{Dove.1986,
-  author   = {Dove, M.T. and Lynden-bell, R.M.},
+  author   = {Dove, M.T. and Lynden-Bell, R.M.},
   doi      = {10.1080/13642818608236861},
   issn     = {1364-2812},
   journal  = {Philosophical Magazine Part B},
