### 2024-01-22 #6150 @flokno
Alright, thank you for the comments! I fixed the bibliography and included the editorial comments in https://github.com/tdep-developers/tdep/commit/9f3e66476745b968520f86797ce288d2760d993f
