+++ paper/paper.md
@@ -84,7 +84,7 @@ In _ab initio_ materials modeling, the contribution of electronic temperature is
 
 An alternative strategy is to construct approximate model Hamiltonians for the nuclear subsystem, which enables an _analytic_ description of the nuclear motion by leveraging perturbation theory starting from an exactly solvable lowest-order model. In this approach, the starting point is given in terms of _phonons_ which are eigensolutions of a Hamiltonian of harmonic form, i.e., quadratic in the nuclear displacements. Anharmonic contributions can be included via established perturbative techniques, in practice up to quartic terms. Higher-order contributions are elusive because the complexity and number of terms in the perturbative expansion grows very quickly with system size and perturbation order. The lattice-dynamics approach is therefore not formally exact. However, the phonon picture is useful for describing a wide range of materials properties in practice, and often reaches excellent accuracy in comparison to experiment while providing precise microscopic insight into the underlying physical phenomena.
 
-The chemical bonding in the lattice dynamics Hamiltonian is represented through force constants. These can be obtained in a purely perturbative, temperature-independent way by constructing a Taylor expansion of the interatomic potential energy about the periodically arranged-atom positions in the crystal. This idea is more than a century old and traces back to Born and von Karman [@Born.1912]. Alternatively, temperature-dependent, _effective_ model Hamiltonians are used in situations where the quadratic term in a bare Taylor expansion is not positive definite, i.e., the average atomic position does _not_ coincide with a minimum of the potential. The classic example is the $^4$He problem, in which a Taylor expansion led to imaginary phonon frequencies in the dense solid phase [@Boer.1948]. Born and coworkers solved this problem in the 1950's by developing a _self-consistent phonon theory_ in which an effective, positive-definite Hamiltonian yielding well-defined phonons is obtained self-consistently using a variational principle [@Born.1951; @Hooton.2010mfn; @Hooton.2010]. An excellent historical review of this development is given in [@Klein.1972].
+The chemical bonding in the lattice dynamics Hamiltonian is represented through force constants. These can be obtained in a purely perturbative, temperature-independent way by constructing a Taylor expansion of the interatomic potential energy about the periodically-arranged atom positions in the crystal. This idea is more than a century old and traces back to Born and von Karman [@Born.1912]. Alternatively, temperature-dependent, _effective_ model Hamiltonians are used in situations where the quadratic term in a bare Taylor expansion is not positive definite, i.e., the average atomic position does _not_ coincide with a minimum of the potential. The classic example is the $^4$He problem, in which a Taylor expansion led to imaginary phonon frequencies in the dense solid phase [@Boer.1948]. Born and coworkers solved this problem in the 1950's by developing a _self-consistent phonon theory_ in which an effective, positive-definite Hamiltonian yielding well-defined phonons is obtained self-consistently using a variational principle [@Born.1951; @Hooton.2010mfn; @Hooton.2010]. An excellent historical review of this development is given in [@Klein.1972].
 
[…]
