+++ paper/paper.md
@@ -122,6 +122,6 @@ We note that parts of the TDEP method have been implemented in other code packag
 
 # Acknowledgements
 
-F.K. and O.H. acknowledge support from the Swedish Research Council (VR) program 2020-04630, and the Swedish e-Science Research Centre (SeRC). Work at Boston College was supported by the U.S. Department of Energy (DOE), Office of Science, Basic Energy Sciences (BES) under Award #DE-SC0021071. M.J.V., A.C., and J.P.B. acknowledge funding by ARC project DREAMS (G.A. 21/25-11) funded by Federation Wallonie Bruxelles and ULiege, and by Excellence of Science project CONNECT number 40007563 funded by FWO and FNRS. R.F. acknowledges financial support by the Spanish State Research Agency under grant number PID2022-139776NB-C62 funded by MCIN/AEI/ 10.13039/501100011033 and by ERDF A way of making Europe. J. K. acknowledges support from the Swedish Research Council (VR) program 2021-00486. I.A.A. and S.I.S. acknowledge support by the Swedish Government Strategic Research Area in Materials Science on Functional Materials at Linköping University (Faculty Grant SFO-Mat-LiU No. 2009 00971). I.A.A. is a Wallenberg Scholar (grant no. KAW-2018.0194). S.I.S. acknowledges the support from Swedish Research Council (VR) (Project No. 2019-05551) and the ERC (synergy grant FASTCORR project 854843).
+F.K. and O.H. acknowledge support from the Swedish Research Council (VR) program 2020-04630, and the Swedish e-Science Research Centre (SeRC). Work at Boston College was supported by the U.S. Department of Energy (DOE), Office of Science, Basic Energy Sciences (BES) under Award #DE-SC0021071. M.J.V., A.C., and J.P.B. acknowledge funding by ARC project DREAMS (G.A. 21/25-11) funded by Federation Wallonie Bruxelles and ULiege, and by Excellence of Science project CONNECT number 40007563 funded by FWO and FNRS. R.F. acknowledges financial support by the Spanish State Research Agency under grant number PID2022-139776NB-C62 funded by MCIN/AEI/ 10.13039/501100011033 and by ERDF A way of making Europe. J. K. acknowledges support from the Swedish Research Council (VR) program 2021-00486. I.A.A. and S.I.S. acknowledge support by the Swedish Government Strategic Research Area in Materials Science on Functional Materials at Linköping University (Faculty Grant SFO-Mat-LiU No. 2009 00971). I.A.A. is a Wallenberg Scholar (grant no. KAW-2018.0194). S.I.S. acknowledges the support from Swedish Research Council (VR) (Project No. 2019-05551) and the ERC (synergy grant FASTCORR project 854843). F.K. acknowledges Christian Carbogno for introducing him to Olle.
 
 # References
