### 2024-01-22 43d7349
extract_forceconstants | correctly dump energies when using >1 MPI rank
