+++ paper/paper.md
@@ -86,7 +86,7 @@ An alternative strategy is to construct approximate model Hamiltonians for the n
 
 The chemical bonding in the lattice dynamics Hamiltonian is represented through force constants. These can be obtained in a purely perturbative, temperature-independent way by constructing a Taylor expansion of the interatomic potential energy about the periodically-arranged atom positions in the crystal. This idea is more than a century old and traces back to Born and von Karman [@Born.1912]. Alternatively, temperature-dependent, _effective_ model Hamiltonians are used in situations where the quadratic term in a bare Taylor expansion is not positive definite, i.e., the average atomic position does _not_ coincide with a minimum of the potential. The classic example is the $^4$He problem, in which a Taylor expansion led to imaginary phonon frequencies in the dense solid phase [@Boer.1948]. Born and coworkers solved this problem in the 1950's by developing a _self-consistent phonon theory_ in which an effective, positive-definite Hamiltonian yielding well-defined phonons is obtained self-consistently using a variational principle [@Born.1951; @Hooton.2010mfn; @Hooton.2010]. An excellent historical review of this development is given in [@Klein.1972].
 
-While the theoretical foundation of (self-consistent) phonon theory has been well-established for decades, more recent developments are concerned with implementing this theory in computer simulations, typically based on density functional theory (DFT) [@Hohenberg.1964; @Kohn.1965]. This has led to a variety of approaches that tackle the self-consistent phonon problem for anharmonic and dynamically stabilized systems [@Souvatzis.2008; @Errea.2013; @Tadano.2015; @Roekeghem.2021; @Monacelli.2021]. Another development was the _non_-self-consistent construction of effective Hamiltonians by optimizing the force constants to the fully anharmonic dynamics observed during MD simulations [@Levy.1984; @Dove.1986]. This idea was extended in the TDEP method to describe phonons in systems like Zirconium in the high-temperature body-centered cubic (bcc) phase, which is dynamically unstable at low temperatures [@Hellman.2011; @Hellman.2013]. While the initial TDEP method was based on _ab initio_ MD, a self-consistent extension was later proposed in the form of _stochastic_ TDEP (sTDEP), where thermal samples are created from the model Hamiltonian itself and the force constants are optimized iteratively until self-consistency [@Shulumba.2017; @Benshalom.2022]. sTDEP furthermore allows to include nuclear quantum effects in materials with light elements in a straightforward way [@Shulumba.20179s8e; @Laniel.2022].
[…]
