### 2024-01-26 #6150 @flokno
Good, this should be it:

There was 1 change in the text (`until self-consistency` -> `until self-consistency is achieved` https://github.com/tdep-developers/tdep/pull/2/commits/42439c53196362e2c67b76030c3a758eb2a58310), one addition in the acknowledgement (https://github.com/tdep-developers/tdep/pull/2/commits/d9706317af24995df2964b14bac7dd53382171a9) and one update in the acknowledgement (https://github.com/tdep-developers/tdep/pull/2/commits/5c7c95de66dc75ca55c6ea201295f2ba77c0d434).

I also addressed @corettialessandro second comment for the docs in the meantime: https://github.com/tdep-developers/tdep/commit/225f60d13eaf3221ef59ac11b96e6957e55236bc

Thanks again everyone for the constructive review!
