### 2024-01-30 9af3949
JOSS Paper and release (#2)
* Paper draft for JOSS

* paper | Intro

* paper | finish intro, add statement of need and summary

* paper | add github action

* paper | minor update

* paper | shorten

* paper | update

* paper | update affiliations

* paper | add Jose and Matthieu

* paper | update authors

* paper | add Johan

* paper | add David + Dennis

* paper | add Matt Heine

* paper | add bib

* paper | bib update

* paper | update bib

* update

* paper | finalize bib

* paper | polish

* paper | polish

* paper | polish

* fix year?

* paper | add Roberta

* paper | add references for
[…]
