### 2024-01-31 #6150 @kyleniemeyer
It works for me - clicking on @rkurchin's name takes me to https://rkurchin.github.io
