### 2024-01-31 #6150 @flokno
Ok it works for me when using Adobe Acrobat, but it does not work using Firefox or Preview (mac).

In Firefox at least, the link points to: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06150/joss.06150/rkurchin.github.io

Should we move on and hope for the best?
